+++ src/pages/search/dashboard/Ecosystems.tsx
@@ -15,7 +15,7 @@ import {
 } from "pages/search/SearchContext";
 import formatNumber from "pages/search/utils/format";
 import matchColor from "pages/search/utils/matchColor";
-import BackendAPI from "pages/search/utils/backendAPI";
+import BackendAPI from "pages/search/api/backendAPI";
 import SmallStackedBar, {
   SmallStackedBarData,
 } from "pages/search/shared_components/charts/SmallStackedBar";
