+++ src/pages/search/dashboard/ecosystems/ecosystemsBox/EcosystemDetails.tsx
@@ -16,7 +16,7 @@ import {
   coverageType,
   SEPADataExt,
 } from "pages/search/types/ecosystems";
-import BackendAPI from "pages/search/utils/backendAPI";
+import BackendAPI from "pages/search/api/backendAPI";
 import SmallStackedBar from "pages/search/shared_components/charts/SmallStackedBar";
 import { wrapperMessage } from "pages/search/types/charts";
 import { EcosystemsController } from "pages/search/dashboard/EcosystemsController";
