+++ src/pages/search/dashboard/landscape/connectivity/CurrentSEPAConnectivity.tsx
@@ -10,7 +10,7 @@ import {
 } from "pages/search/SearchContext";
 import formatNumber from "pages/search/utils/format";
 import matchColor from "pages/search/utils/matchColor";
-import BackendAPI from "pages/search/utils/backendAPI";
+import BackendAPI from "pages/search/api/backendAPI";
 import TextBoxes from "pages/search/shared_components/TextBoxes";
 
 import {
