+++ src/pages/search/dashboard/landscape/forest/ForestIntegrity.tsx
@@ -11,7 +11,7 @@ import TextBoxes from "pages/search/shared_components/TextBoxes";
 import { IconTooltip } from "pages/search/shared_components/Tooltips";
 
 import matchColor from "pages/search/utils/matchColor";
-import BackendAPI from "pages/search/utils/backendAPI";
+import BackendAPI from "pages/search/api/backendAPI";
 
 import {
   SearchLegacyCTX,
