+++ src/pages/search/dashboard/landscape/humanFootprint/TimelineFootprintController.ts
@@ -1,5 +1,5 @@
 import formatNumber from "pages/search/utils/format";
-import BackendAPI from "pages/search/utils/backendAPI";
+import BackendAPI from "pages/search/api/backendAPI";
 import { ShapeLayer } from "pages/search/types/layers";
 import matchColor from "pages/search/utils/matchColor";
 import { ShapeAPIObject } from "pages/search/types/api";
